+++ splayout/adjointmethod/topologyregion3d.py
@@ -67,7 +67,8 @@ class TopologyOptRegion3D:
         self.rename = rename
         self.index_region_name = self.rename + "_index"
         self.field_region_name = self.rename + "_field"
-        self.__initialize()
+        if not(self.fdtd_engine is None):
+            self.__initialize()
         self.epsilon_figure = None
         self.field_figure = None