+++ splayout/components/__init__.py
@@ -18,4 +18,4 @@ from .slowlyvaryingtaper import SlowlyVaryingTaper
 from .text import Text
 from .waveguide import Waveguide, ArbitraryAngleWaveguide
 from .sbend import SBend,ASBend
-from .filledpattern import Circle,Rectangle
+from .filledpattern import Circle, Rectangle